fix: theme guard
