+++ packages/core/src/core/providers/index.ts
@@ -1,0 +1,1 @@
+export * from './default-theme.provider';
