+++ packages/startup/src/common/constants.ts
@@ -1,0 +1,9 @@
+export const overrideColorTokens = {
+  'editorGroupHeader.tabsBackground': '#ECF1FE',
+  'editor.background': '#fff',
+  'aiNative.inlineDiffAddedRange': '#26bf6d1f',
+  'aiNative.inlineDiffRemovedRange': '#ff4d4f1e',
+  'aiNative.inlineDiffAcceptPartialEdit': '#26bf6d80',
+  'aiNative.inlineDiffDiscardPartialEdit': '#ff4d4f80',
+  'aiNative.inlineDiffAcceptPartialEdit.foreground': '#000',
+}
